Align dependency layout with webpack
