+++ crates/unpack_core/src/dependencies/harmony_export_expression_dependency.rs
@@ -1,0 +1,24 @@
+use serde::{Deserialize, Serialize};
+
+use crate::SourceRange;
+
+#[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
+pub struct HarmonyExportExpressionDependency {
+    pub range: SourceRange,
+    pub statement_range: SourceRange,
+    pub declaration_id: Option<String>,
+}
+
+impl HarmonyExportExpressionDependency {
+    pub fn new(
+        range: SourceRange,
+        statement_range: SourceRange,
+        declaration_id: Option<String>,
+    ) -> Self {
+        Self {
+            range,
+            statement_range,
+            declaration_id,
+        }
+    }
+}
