+++ crates/unpack_core/src/dependencies/harmony_export_specifier_dependency.rs
@@ -1,0 +1,16 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
+pub struct HarmonyExportSpecifierDependency {
+    pub id: String,
+    pub name: String,
+}
+
+impl HarmonyExportSpecifierDependency {
+    pub fn new(id: impl Into<String>, name: impl Into<String>) -> Self {
+        Self {
+            id: id.into(),
+            name: name.into(),
+        }
+    }
+}
