+++ crates/unpack_core/src/dependencies/import_dependency.rs
@@ -1,0 +1,27 @@
+use serde::{Deserialize, Serialize};
+
+use super::ModuleDependency;
+use crate::SourceRange;
+
+#[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
+pub struct ImportDependency {
+    pub module: ModuleDependency,
+}
+
+impl ImportDependency {
+    pub fn new(
+        request: impl Into<String>,
+        range: SourceRange,
+        source_order: Option<usize>,
+    ) -> Self {
+        let mut module = ModuleDependency::new(request, source_order);
+        module.range = Some(range);
+        Self { module }
+    }
+
+    pub fn range(&self) -> SourceRange {
+        self.module
+            .range
+            .expect("import dependency should have range")
+    }
+}
