+++ crates/unpack_core/src/dependencies/null_dependency.rs
@@ -1,0 +1,4 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
+pub struct NullDependency;
