+++ crates/unpack_core/src/lib.rs
@@ -1,3 +1,4 @@
+mod async_dependencies_block;
 mod build_cache;
 mod build_chunk_graph;
 mod cache_hash;
@@ -8,6 +9,8 @@ mod code_generation;
 mod code_generation_record;
 mod compilation;
 mod compiler;
+mod dependencies;
+mod dependencies_block;
 mod dependency;
 mod error;
 mod exports_info;
@@ -20,6 +23,7 @@ mod logging;
 mod make;
 mod module;
 mod module_graph;
+mod module_graph_connection;
 mod normal_module_factory;
 mod output_filename;
 mod pack_file;
@@ -29,6 +33,7 @@ mod resolver;
 mod runtime;
 mod snapshot;
 
+pub use async_dependencies_block::AsyncDependenciesBlock;
 pub use build_cache::{BuildDependency, CacheCompression, CacheKind, CacheOptions};
 pub use chunk::{Chunk, ChunkHandle};
 pub use chunk_graph::ChunkGraph;
@@ -39,23 +44,22 @@ pub use compiler::{
     CacheIdleReason, CacheLifecycleOutcome, Compiler, CompilerOptions, DEFAULT_EXTENSIONS, Entry,
     PendingCompilation,
 };
-pub use dependency::{
-    AsyncDependenciesBlock, ConstDependency, Dependency, DependencyKind, EntryDependency,
-    HarmonyExportExpressionDependency, HarmonyExportHeaderDependency,
-    HarmonyExportImportedSpecifierDependency, HarmonyExportSpecifierDependency,
-    HarmonyImportSideEffectDependency, HarmonyImportSpecifierDependency, ImportDependency,
-    ModuleDependency, NullDependency, SourceRange,
+pub use dependencies::{
+    ConstDependency, EntryDependency, HarmonyExportExpressionDependency,
+    HarmonyExportHeaderDependency, HarmonyExportImportedSpecifierDependency,
+    HarmonyExportSpecifierDependency, HarmonyImportSideEffectDependency,
+    HarmonyImportSpecifierDependency, ImportDependency, ModuleDependency, NullDependency,
 };
+pub use dependencies_block::DependenciesBlock;
+pub use dependency::{Dependency, DependencyKind, SourceRange};
 pub use error::{Error, Result};
 pub use exports_info::ExportsInfo;
 pub use hooks::{CompilationHooks, HookFuture};
 pub use loader::{LoaderFuture, LoaderRequest, LoaderRunner, MatchedLoader, ModuleRule};
 pub use logging::{InfrastructureLogEvent, InfrastructureLogLevel, InfrastructureLoggingOptions};
 pub use module::{Module, ModuleHandle, ModuleIdentity, ModuleType};
-pub use module_graph::{
-    AsyncDependenciesBlockIndex, DependencyIndex, ModuleGraph, ModuleGraphConnection,
-    ModuleGraphConnectionHandle,
-};
+pub use module_graph::{AsyncDependenciesBlockIndex, DependencyIndex, ModuleGraph};
+pub use module_graph_connection::{ModuleGraphConnection, ModuleGraphConnectionHandle};
 pub use normal_module_factory::{FactorizedModule, NormalModuleFactory};
 pub use resolver::{ResolveOptions, ResolvedResource, UnpackResolver};
 pub use snapshot::{SnapshotOptions, SnapshotPathPattern, SnapshotStrategy};
